who needs strategy
