--- src/org/thoughtcrime/securesms/components/camera/PreviewStrategy.java
@@ -1,5 +1,0 @@
-package org.thoughtcrime.securesms.components.camera;
-
-public interface PreviewStrategy extends com.commonsware.cwac.camera.PreviewStrategy {
-  boolean isReady();
-}
